fixing auth
the cookie solution won't be a reliable thing in the end, since I noted that we will be getting lots of CORS issues, and even in browsers that this is still working, it is very likelly that it will stop on a near future. So Im using a "webhook" approach , passing back the token to the previous service on a specific url. then it's reponsability of that service to do what it needs (if it needs) on its own end.